BackEnd/requirements.txt
flask
flask-cors
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
shap
joblib
Pillow